silversmith modify

+++ silversmith/silversmith.sql
@@ -26,7 +26,7 @@ CREATE TABLE `tb_character` (
   `viptype` tinyint(4) DEFAULT '0' COMMENT '角色的类型（0普通 1 VIP1  2 vip2）',
   `nickname` varchar(255) COLLATE utf8_unicode_ci NOT NULL DEFAULT '' COMMENT '角色的昵称',
   `profession` tinyint(4) unsigned zerofill NOT NULL DEFAULT '0000' COMMENT '角色的职业 （0 新手 1战士 2 法师 3 游侠 4 牧师）',
-  `figure` tinyint(4) DEFAULT '0' COMMENT '角色的形象',
+  `figure` int(10) DEFAULT '0' COMMENT '角色的形象',
   `sex` tinyint(4) DEFAULT '1' COMMENT '1男 2女',
   `camp` tinyint(4) DEFAULT '0' COMMENT '所属阵营 0无 1魏 2蜀 3 吴',
   `level` int(10) DEFAULT '1' COMMENT '角色的等级 初始为1',
